requirements.txt
blinker==1.8.2
certifi==2025.7.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.0.3
Flask-Mail==0.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
midtransclient==1.4.2
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
PyJWT==2.9.0
pymongo==4.13.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==65.5.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
gunicorn==21.2.0
cryptography==45.0.5
cffi==1.17.1
pycparser==2.22
